Major changes/fixes/organization
## Changes:

- Script variables are now only used when necessary: $cdlSettings and $cdlLockers
- Functions that depended on the script variables now have parameters to take them in (Add-LockerMetadata, Invoke-PasswordHashing, etc.
- Reworked some of the mechanics to reduce repeated code
- Renamed variables to better reflect what they are
- Changed Write-Outputs back to Write-Hosts to use ForegroundColors
- Moved functions and sections around to make the program properly execute and reflect logical organization

## Debugging:

FileSystemWatcher needs to be troubleshooted further - it isn't actively refreshing the arrays ($cdlSettings and $cdlLockers) when they are manually updated. However, I still get prompted with the update window each time the program starts up, it is a bit annoying -- may want to not display the window unless there is a newer version.

In short, I want to say that the settings are loaded in but the lockers are not.